% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_subtitles.R
\name{as_tibble.Subtitles}
\alias{as_tibble.Subtitles}
\title{Convert subtitles to a tibble}
\usage{
\method{as_tibble}{Subtitles}(x)
}
\arguments{
\item{x}{an object of class \code{Subtitles} or \code{MultiSubtitles}.}
}
\value{
A \code{tibble}.
}
\description{
This function converts \code{Subtitles} and \code{MultiSubtitles} objects
to a \code{tibble} with one subtitle line per row; text, timecodes,
and (expanded) meta-data as columns.
}
