% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_subtitles.R
\name{bind_subtitles}
\alias{bind_subtitles}
\title{Bind subtitles}
\usage{
bind_subtitles(..., collapse = TRUE, sequential = TRUE)
}
\arguments{
\item{...}{\code{subtitles} or \code{multisubtitles} objects to be binded.}

\item{collapse}{logical. If \code{TRUE}, subtitles are combined in a single \code{subtitles} object.}

\item{sequential}{logical. If \code{TRUE} (default) timecodes
are recalculated to follow concatenation.}
}
\value{
A \code{subtitles} object if \code{collapse = TRUE} (default).
A \code{multisubtitles} object if \code{collapse = FALSE}.
}
\description{
Bind \code{subtitles} or/and \code{multisubtitles} objects.
}
\examples{
f1 <- system.file("extdata", "ex_subrip.srt", package = "subtools")
s1 <- read_subtitles(f1, metadata = tibble::tibble(Season = 1, Episode = 2))
f2 <- system.file("extdata", "ex_substation.ass", package = "subtools")
s2 <- read_subtitles(f2, metadata = tibble::tibble(Season = 2))
bind_subtitles(s1, s2)
bind_subtitles(s1, s2, collapse = FALSE)
}
