% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subtitles.R
\name{clean_tags}
\alias{clean_tags}
\alias{clean_captions}
\alias{clean_patterns}
\title{Clean subtitles}
\usage{
clean_tags(x, format = "srt", clean.empty = TRUE)

clean_captions(x, clean.empty = TRUE)

clean_patterns(x, pattern, clean.empty = TRUE)
}
\arguments{
\item{x}{a \code{subtitles} or \code{multisubtitles} object.}

\item{format}{the original format of the \code{subtitles} objects.}

\item{clean.empty}{logical. Should empty remaining lines ("") deleted after cleaning.}

\item{pattern}{a character string containing a regular expression to be matched and cleaned.}
}
\value{
A \code{subtitles} or \code{multisubtitles} object.
}
\description{
Functions to clean subtitles. \code{clean_tags} cleans formatting tags.
\code{clean_captions} cleans close captions, i.e all text enclosed in parentheses or squared brackets.
\code{clean_patterns} provides a more general and flexible cleaning based on regular expressions.
}
