% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subtitles.R
\name{cleanTags}
\alias{cleanTags}
\title{Title}
\usage{
cleanTags(x, format = "srt", clean.empty = TRUE)
}
\arguments{
\item{x}{xxx}

\item{format}{xxx}

\item{clean.empty}{xxx}
}
\value{
xxx
}
\description{
Title
}

