% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_subtitles.R
\name{combineSubs}
\alias{combineSubs}
\title{Title}
\usage{
combineSubs(..., collapse = TRUE, sequential = TRUE)
}
\arguments{
\item{...}{\code{Subtitles} or \code{MultiSubtitles} objects to be concatenated.}

\item{collapse}{XXX}

\item{sequential}{logical. If \code{TRUE} (default) timecodes
are recalculated to follow concatenation.}
}
\value{
xxx
}
\description{
Title
}

