% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_subtitles.R
\name{get_raw_text}
\alias{get_raw_text}
\title{Get subtitles text}
\usage{
get_raw_text(x, collapse = " ")
}
\arguments{
\item{x}{an object of class \code{subtitles} or \code{multisubtitles}.}

\item{collapse}{a character string to separate the subtitles lines.}
}
\value{
A character string.
}
\description{
This function extracts the raw text content of subtitles objects as a character string.
}
\examples{
f <- system.file("extdata", "ex_subrip.srt", package = "subtools")
s <- read_subtitles(f)
get_raw_text(s)
cat(get_raw_text(s, collapse = "\\n"))

}
