% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{get_subtitles_info}
\alias{get_subtitles_info}
\title{Get basic informations for subtitle objects}
\usage{
get_subtitles_info(x)
}
\arguments{
\item{x}{a \code{subtitles} or \code{multisubtitles} object.}
}
\value{
Called for its side effects (printing to the console). Returns \code{NULL} invisibly.
}
\description{
Get basic informations for subtitle objects
}
\examples{
s <- read_subtitles(
  system.file("extdata", "ex_subrip.srt", package = "subtools")
)
get_subtitles_info(s)

}
