% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mkv.R
\name{mkvInfo}
\alias{mkvInfo}
\title{Get informations about subtitles embedded in MKV files}
\usage{
mkvInfo(file, mkvmerge.exec = "mkvmerge", print.info = TRUE)
}
\arguments{
\item{file}{a character string giving the path to the MKV file.}

\item{mkvmerge.exec}{a character string giving the path to the \code{mkvmerge} executable.}

\item{print.info}{print basic informations about subtitle tracks. Default is \code{TRUE}.}
}
\value{
A list with complete data about the MKV is invisibly returned.
If the MKV has at least 1 subtitles track and \code{print.info} is \code{TRUE}, basic informations are printed.
Otherwise it returns a warning.
}
\description{
This function uses \code{mkvmerge} to extract tracks data from an MKV file.
You must have \code{mkvmerge} installed on your computer.
}
\references{
\url{https://mkvtoolnix.download/downloads.html}
}
