% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mkv.R
\name{mkvSubExtract}
\alias{mkvSubExtract}
\title{Extract subtitles embedded in MKV files}
\usage{
mkvSubExtract(file, id = 2, mkvextract.exec = "mkvextract",
  mkvmerge.exec = "mkvmerge")
}
\arguments{
\item{file}{a character string giving the path to the MKV file.}

\item{id}{An integer giving the ID of the tracks to be extracted.
Can be a vector of length > 1 to extract several tracks.
If \code{NA} (default), the default subtitle tracks will be extracted}

\item{mkvextract.exec}{a character string giving the path to the \code{mkvextract} executable.}

\item{mkvmerge.exec}{character string giving the path to the \code{mkvmerge} executable.}
}
\value{
An object of class \code{Subtitles} (see \code{\link{Subtitles}}).
If several tracks are requested (via \code{id}), an object of class \code{MultiSubtitles};
i.e. a list of \code{\link{Subtitles}} objects.
}
\description{
This function uses \code{mkvextract} to extract subtitles from an MKV file.
You must have \code{mkvmerge} and \code{mkvextract} installed on your computer.
}
\details{
The function \code{\link{mkvInfo}} is a simple way to identify the ID of subtitles tracks from a MKV file.

Not all the subtitle formats supported by \code{mkvextract} can be read by \code{subtools}.
See \code{\link{read_subtitles}} for the list of formats currently supported.
}
\references{
\url{https://mkvtoolnix.download/downloads.html}
}
