% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_subtitles.R
\name{move_subtitles}
\alias{move_subtitles}
\title{Move subtitles
Move subtitles forward or backward \code{subtitles}.}
\usage{
move_subtitles(x, lag)
}
\arguments{
\item{x}{\code{subtitles} or \code{multisubtitles} objects to be delayed.}

\item{lag}{\code{numeric.} Number of seconds the subtitles should be moved,
forward if positive and backward if negative.}
}
\value{
A \code{subtitles} or \code{multisubtitles} object.
}
\description{
Move subtitles
Move subtitles forward or backward \code{subtitles}.
}
\examples{
  f1 <- system.file("extdata", "ex_subrip.srt", package = "subtools")
  s1 <- read_subtitles(f1, metadata = tibble::tibble(Season = 1, Episode = 2))
  s2 <- move_subtitles(s1, 2.7)
}
