% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{print.Subtitles}
\alias{print.Subtitles}
\title{Print \code{Subtitles} objects}
\usage{
\method{print}{Subtitles}(x, printlen = 1000L)
}
\arguments{
\item{x}{a \code{Subtitles} object.}

\item{printlen}{the maximum number of subtitles to print.}
}
\description{
Print \code{Subtitles} objects
}

