% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{print.Subtitles}
\alias{print.Subtitles}
\alias{print.MultiSubtitles}
\title{Print methods for subtitles}
\usage{
\method{print}{Subtitles}(x, printlen = 1000L, ...)

\method{print}{MultiSubtitles}(x, printlen = 10L, ...)
}
\arguments{
\item{x}{a \code{Subtitles} or \code{MultiSubtitles} object.}

\item{printlen}{the maximum number of subtitles to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print methods for subtitles
}
