% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{read.subtitles}
\alias{read.subtitles}
\title{Read subtitles}
\usage{
read.subtitles(file, format = "auto", clean.tags = TRUE,
  metadata = list())
}
\arguments{
\item{file}{the name of the file which the subtitles are to be read from.
If it does not contain an absolute path, the file name is relative to the current working directory.}

\item{format}{a character string specifying the format of the subtitles.
Five formats can be read: \code{"subrip"}, \code{"substation"}, \code{"microdvd"}, \code{"subviewer"} and \code{"webvtt"}.
Default is \code{"auto"} which tries to detect automatically the format of the file from its extension.}

\item{clean.tags}{logical. If \code{"TRUE"}, formating tags are deleted from subtitles using \code{\link{cleanTags}}.}

\item{metadata}{a named list of metadata to be attached to the subtitles.}
}
\value{
An object of class \code{Subtitles} (see \code{\link{Subtitles}}).
}
\description{
Reads subtitles from a file.
}

