% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_series.R
\name{read.subtitles.season}
\alias{read.subtitles.season}
\title{Read series subtitles}
\usage{
read.subtitles.season(dir, quietly = FALSE, format = "auto", ...)
}
\arguments{
\item{dir}{the name of the directory which the subtitles are to be read from (see Details).}

\item{quietly}{a logical. If \code{FALSE} (default), a message indicating the number of imported files is printed.}
}
\value{
An object of class \code{MultiSubtitles};
i.e. a list of \code{\link{Subtitles}} objects.
}
\description{
These functions read one or several subtitles files organized in directories.
They are especially designed to import subtitles of series with multiple episodes.
}
\details{
Series/Series_1/Season_1
'\preformatted{This text stays as is
'and will
'not be wrapped.}
}

