% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_series.R
\name{read.subtitles.season}
\alias{read.subtitles.multiseries}
\alias{read.subtitles.season}
\alias{read.subtitles.serie}
\title{Read series subtitles}
\usage{
read.subtitles.season(dir, quietly = FALSE, format = "auto", ...)

read.subtitles.serie(dir, quietly = FALSE, format = "auto", ...)

read.subtitles.multiseries(dir, quietly = FALSE, format = "auto", ...)
}
\arguments{
\item{dir}{the name of the directory which the subtitles are to be read from (see Details).}

\item{quietly}{a logical. If \code{FALSE} (default), a message indicating the number of imported files is printed.}

\item{format}{a character string specifying the format of the subtitles
(default is "\code{auto}", see \code{\link{read.subtitles}} for details).}

\item{...}{further arguments to be passed to \code{\link{read.subtitles}}.}
}
\value{
An object of class \code{MultiSubtitles};
i.e. a list of \code{\link{Subtitles}} objects.
}
\description{
These functions read one or several subtitles files organized in directories.
They are especially designed to import subtitles of series with multiple episodes.
}
\details{
These functions read subtitles files at different levels from a 3-levels directory (see the tree below).
The function \code{read.subtitles.multiseries} reads everything recursively from "Series_Collection".
The function \code{read.subtitles.serie} reads everything recursively from a serie folder (e.g. "Serie_A").
The function \code{read.subtitles.season} reads everything from a season folder (e.g. "Season_1").
To read a specific episode file (e.g. "Episode_1.srt), use \code{\link{read.subtitles}}.
\preformatted{
Series_Collection
|-- Serie_A
|   |-- Season_1
|   |   |-- Episode_1.srt
|-- Serie_B
|   |-- Season_1
|   |   |-- Episode_1.srt
|   |-- Season_2
|   |   |-- Episode_1.srt
|   |   |-- Episode_2.srt}
}

