% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{read_subtitles}
\alias{read_subtitles}
\title{Read subtitles}
\usage{
read_subtitles(
  file,
  format = "auto",
  clean.tags = TRUE,
  metadata = data.frame(),
  frame.rate = NA,
  encoding = "auto"
)
}
\arguments{
\item{file}{a file path, URL, or literal text. If a file path does not
contain an absolute path, it is relative to the current working directory.
Literal text can be a multi-line string or a character vector.}

\item{format}{a character string specifying the format of the subtitles.
Four formats can be read: \code{"subrip"}, \code{"substation"}, \code{"microdvd"}, \code{"subviewer"} (v.2) and \code{"webvtt"}.
Default is \code{"auto"} which tries to detect automatically the format of the file from its extension or its content.}

\item{clean.tags}{logical. If \code{"TRUE"} (default), formatting tags are deleted from subtitles using \code{\link{clean_tags}}.}

\item{metadata}{a one-row dataframe or tibble, or any object that can be coerced
into a one-row tibble by \code{link[tibble]{as_tibble}}.}

\item{frame.rate}{a numeric value giving the frame rate in frames per second. Only relevant for MicroDVD format.
If \code{NA} (default), the function tries to extract the frame.rate from the file.
If it fails, the frame rate is set at 24p (23.976).}

\item{encoding}{the name of the encoding to be used. Default is "\code{auto}" and
uses \code{\link[readr]{guess_encoding}} to detect encoding.}
}
\value{
An object of class \code{subtitles} (see \code{\link{subtitles}}).
}
\description{
Reads subtitles from a file, URL, or text.
}
\details{
The support of WebVTT is basic and experimental.
}
\examples{

# read a SubRip file
f <- system.file("extdata", "ex_subrip.srt", package = "subtools")
f <- system.file("extdata", "ex_webvtt.vtt", package = "subtools")
read_subtitles(f)

# read literal subtitle text
read_subtitles("WEBVTT\n\n00:00:01.000 --> 00:00:02.000\nHello")

}
