% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subtitles.R
\name{sentencify}
\alias{sentencify}
\title{Reorganize subtitles as sentences}
\usage{
sentencify(x)
}
\arguments{
\item{x}{a \code{Subtitles} object.}
}
\value{
A \code{Subtitles} object.
}
\description{
This function reorganizes a \code{Subtitles} object
in order that each subtitle line is a complete sentence.
}

