% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{[.Subtitles}
\alias{[.Subtitles}
\title{Extract parts of \code{Subtitles} objects}
\usage{
\method{[}{Subtitles}(x, i)
}
\arguments{
\item{x}{a \code{Subtitles} object.}

\item{i}{elements to extract.
Can be numeric, character, or logical.}
}
\value{
A \code{Subtitles} object.
}
\description{
Extract parts of \code{Subtitles} objects
}

