% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{sub_info}
\alias{sub_info}
\title{Get basic informations for subtitle objects}
\usage{
sub_info(x)
}
\arguments{
\item{x}{a \code{Subtitles} or \code{MultiSubtitles} object.}
}
\description{
Get basic informations for subtitle objects
}
\examples{
s <- read_subtitles(
  system.file("extdata", "ex_subrip.srt", package = "subtools")
)
sub_info(s)

}
