% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_subtitles.R
\name{subDataFrame}
\alias{subDataFrame}
\title{Convert subtitles to a dataframe}
\usage{
subDataFrame(x)
}
\arguments{
\item{x}{an object of class \code{Subtitles} or \code{MultiSubtitles}.}
}
\value{
A data frame.
}
\description{
This function converts \code{Subtitles} and \code{MultiSubtitles} objects
to a dataframe with one subtitle line per row; text, timecodes,
and (expanded) meta-data as columns.
}

