% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{summary.Subtitles}
\alias{summary.MultiSubtitles}
\alias{summary.Subtitles}
\title{Summary methods for subtitles}
\usage{
\method{summary}{Subtitles}(x, ...)

\method{summary}{MultiSubtitles}(x, ...)
}
\arguments{
\item{x}{a \code{Subtitles} or \code{MultiSubtitles} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary methods for subtitles
}

