% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_subtitles.R
\name{tmCorpus}
\alias{tmCorpus}
\title{Convert subtitles to a tm corpus}
\usage{
tmCorpus(x, collapse = " ")
}
\arguments{
\item{x}{an object of class \code{Subtitles} or \code{MultiSubtitles}.}

\item{collapse}{a character string to separate the subtitles lines.}
}
\value{
An object of class \code{VCorpus}.
}
\description{
This function converts \code{Subtitles} and \code{MultiSubtitles} objects
to virtual corpora as defined by package \pkg{tm}.
}
\seealso{
\code{\link[tm]{Corpus}}, \code{\link[tm]{VCorpus}}.
}

