% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_subtitles.R
\name{write_subtitles}
\alias{write_subtitles}
\title{Write subtitles}
\usage{
write_subtitles(x, file, format = "srt", encoding = "UTF-8")
}
\arguments{
\item{x}{an object of class \code{subtitles}.}

\item{file}{a character string naming a file for writing.}

\item{format}{a character string giving the file format.
Not used (only SubRip format is currently implemented).}

\item{encoding}{the name of the encoding to be used.}
}
\value{
Called for its side effects (writing to \code{file}). Returns \code{NULL} invisibly.
}
\description{
This function writes a \code{subtitles} object in a file.
}
