% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{as_subtitle}
\alias{as_subtitle}
\alias{as_subtitle.default}
\alias{as_subtitle.character}
\title{Convert an R object to a \code{subtitles} object}
\usage{
as_subtitle(
  x,
  format = "auto",
  clean.tags = TRUE,
  metadata = data.frame(),
  frame.rate = NA,
  encoding = "auto",
  ...
)

\method{as_subtitle}{default}(
  x,
  format = "auto",
  clean.tags = TRUE,
  metadata = data.frame(),
  frame.rate = NA,
  encoding = "auto",
  ...
)

\method{as_subtitle}{character}(
  x,
  format = "auto",
  clean.tags = TRUE,
  metadata = data.frame(),
  frame.rate = NA,
  encoding = "auto",
  ...
)
}
\arguments{
\item{x}{an R object that can be coerced into a \code{subtitles} object}

\item{format}{a character string specifying the format of the subtitles.
Four formats can be read: \code{"subrip"}, \code{"substation"}, \code{"microdvd"}, \code{"subviewer"} (v.2) and \code{"webvtt"}.
Default is \code{"auto"} which tries to detect automatically the format of the file from its extension or its content.}

\item{clean.tags}{logical. If \code{"TRUE"} (default), formatting tags are deleted from subtitles using \code{\link{clean_tags}}.}

\item{metadata}{a one-row dataframe or tibble, or any object that can be coerced
into a one-row tibble by \code{link[tibble]{as_tibble}}.}

\item{frame.rate}{a numeric value giving the frame rate in frames per second. Only relevant for MicroDVD format.
If \code{NA} (default), the function tries to extract the frame.rate from the file.
If it fails, the frame rate is set at 24p (23.976).}

\item{encoding}{the name of the encoding to be used. Default is "\code{auto}" and
uses \code{\link[readr]{guess_encoding}} to detect encoding.}

\item{...}{passed on to downstream methods.}
}
\description{
Convert an R object to a \code{subtitles} object
}
\examples{
as_subtitle(
  c("WEBVTT",
    "X-TIMESTAMP-MAP=MPEGTS:181083,LOCAL:00:00:00.000",
    "",
    "3",
    "00:00:21.199 --> 00:00:22.333", ">> FEMALE SPEAKER:",
    "Don't stay up too late.",
    "",
    ""
  ), format = "webvtt"
)
}
