% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{as_subtitle}
\alias{as_subtitle}
\alias{as_subtitle.default}
\alias{as_subtitle.character}
\title{Convert an R object to a \code{Subtitles} object}
\usage{
as_subtitle(x, format = "auto", clean.tags = TRUE, metadata = list(),
  frame.rate = NA, encoding = "UTF-8", ...)

\method{as_subtitle}{default}(x, format = "auto", clean.tags = TRUE,
  metadata = list(), frame.rate = NA, encoding = "UTF-8", ...)

\method{as_subtitle}{character}(x, format = "auto", clean.tags = TRUE,
  metadata = list(), frame.rate = NA, encoding = "UTF-8", ...)
}
\arguments{
\item{format}{a character string specifying the format of the subtitles.
Four formats can be read: \code{"subrip"}, \code{"substation"}, \code{"microdvd"}, \code{"subviewer"} (v.2) and \code{"webvtt"}.
Default is \code{"auto"} which tries to detect automatically the format of the file from its extension.}

\item{clean.tags}{logical. If \code{"TRUE"} (default), formating tags are deleted from subtitles using \code{\link{cleanTags}}.}

\item{metadata}{a named list of metadata to be attached to the subtitles.}

\item{frame.rate}{a numeric value giving the frame rate in frames per second. Only relevant for MicroDVD format.
If \code{NA} (default), the function tries to extract the frame.rate from the file.
If it fails, the frame rate is set at 24p (23.976).}

\item{encoding}{the name of the encoding to be used.}

\item{...}{passed on to downstream methods}

\item{an}{R object that can be coerced into a \code{Subtitles} object}
}
\description{
Convert an R object to a \code{Subtitles} object
}
\examples{
as_subtitle(
  c("WEBVTT",
    "X-TIMESTAMP-MAP=MPEGTS:181083,LOCAL:00:00:00.000",
    "",
    "3",
    "00:00:21.199 --> 00:00:22.333", ">> FEMALE SPEAKER:",
    "Don't stay up too late.",
    "",
    ""
  ), format = "webvtt"
)
}
