% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subtitles.R
\name{cleanTags}
\alias{cleanCaptions}
\alias{cleanPatterns}
\alias{cleanTags}
\title{Clean subtitles}
\usage{
cleanTags(x, format = "srt", clean.empty = TRUE)

cleanCaptions(x, clean.empty = TRUE)

cleanPatterns(x, pattern, clean.empty = TRUE)
}
\arguments{
\item{x}{a \code{Subtitles} object.}

\item{format}{the original format of the \code{Subtitles} object.}

\item{clean.empty}{logical. Should empty remaining lines ("") deleted after cleaning.}

\item{a}{character string containing a regular expression to be matched and clean.}
}
\value{
A \code{Subtitles} object.
}
\description{
Functions to clean subtitles. \code{cleanTags} cleans formatting tags.
\code{cleanCaptions} cleans close captions.
\code{cleanPatterns} provide a more general and flexible cleaning function base on regular expressions.
}

