% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_subtitles.R
\name{combineSubs}
\alias{combineSubs}
\title{Combine subtitles}
\usage{
combineSubs(..., collapse = TRUE, sequential = TRUE)
}
\arguments{
\item{...}{\code{Subtitles} or \code{MultiSubtitles} objects to be concatenated.}

\item{collapse}{logical. If \code{TRUE}, subtitles are combined in a single \code{Subtitles} object.}

\item{sequential}{logical. If \code{TRUE} (default) timecodes
are recalculated to follow concatenation.}
}
\value{
A \code{Subtitles} object if \code{collapse = TRUE} (default).
A \code{MultiSubtitles} object if \code{collapse = FALSE}.
}
\description{
Combine \code{Subtitles} or/and \code{MultiSubtitles} objects.
}

