% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_subtitles.R
\name{combineSubs}
\alias{combineSubs}
\title{Combine subtitles}
\usage{
combineSubs(..., collapse = TRUE, sequential = TRUE)
}
\arguments{
\item{...}{\code{Subtitles} or \code{MultiSubtitles} objects to be concatenated.}

\item{collapse}{logical. If \code{TRUE}, subtitles are combined in a single \code{Subtitles} object.}

\item{sequential}{logical. If \code{TRUE} (default) timecodes
are recalculated to follow concatenation.}
}
\value{
A \code{Subtitles} object if \code{collapse = TRUE} (default).
A \code{MultiSubtitles} object if \code{collapse = FALSE}.
}
\description{
Combine \code{Subtitles} or/and \code{MultiSubtitles} objects.
}
\examples{
f1 <- system.file("extdata", "ex_subrip.srt", package = "subtools")
s1 <- read.subtitles(f1)
f2 <- system.file("extdata", "ex_substation.ass", package = "subtools")
s2 <- read.subtitles(f2)
combineSubs(s1, s2)

}

