% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_subtitles.R
\name{combineSubs}
\alias{combineSubs}
\title{Title}
\usage{
combineSubs(..., collapse = TRUE, sequential = TRUE)
}
\arguments{
\item{...}{xxx}

\item{collapse}{xxx}

\item{sequential}{xxx}
}
\value{
xxx
}
\description{
Title
}

