% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{print.multisubtitles}
\alias{print.multisubtitles}
\title{Print method for multisubtitles}
\usage{
\method{print}{multisubtitles}(x, printlen = 10L, ...)
}
\arguments{
\item{x}{a \code{multisubtitles} object.}

\item{printlen}{the maximum number of subtitles to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for multisubtitles
}
\examples{
f <- system.file("extdata", "ex_subrip.srt", package = "subtools")
s <- read_subtitles(f)
bind_subtitles(s, s, collapse = FALSE)

}
