% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{print.MultiSubtitles}
\alias{print.MultiSubtitles}
\title{Print method for MultiSubtitles}
\usage{
\method{print}{MultiSubtitles}(x, printlen = 10L, ...)
}
\arguments{
\item{x}{a \code{MultiSubtitles} object.}

\item{printlen}{the maximum number of subtitles to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for MultiSubtitles
}
\examples{
f <- system.file("extdata", "ex_subrip.srt", package = "subtools")
s <- read_subtitles(f)
bind_subs(s, s, collapse = FALSE)

}
