% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subtitles.R
\name{Subtitles}
\alias{Subtitles}
\title{Create a \code{Subtitles} object}
\usage{
Subtitles(text, timecode.in, timecode.out, id, metadata = data.frame())
}
\arguments{
\item{text}{a character vector of subtitles text content.}

\item{timecode.in}{a character vector giving the time that the subtitles appear on the screen.
The format must be "HH:MM:SS.mS".}

\item{timecode.out}{a character vector giving the time that the subtitles disappear.
The format must be "HH:MM:SS.mS".}

\item{id}{a vector of numeric ID for subtitles.
If not provided it is generated automatically from \code{timecode.in} order.}

\item{metadata}{a one-row dataframe or tibble, or any object that can be coerced
into a one-row tibble by \code{link[tibble]{as_tibble}}.}
}
\value{
a \code{Subtitles} object i.e. a \code{tibble} with at least 4 columns containing IDs,
timecodes and text of the subtitles and optionally metadata in extra columns.
}
\description{
A \code{Subtitles} is a special form of \code{tibble}.
}
