% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_subtitles.R
\name{write.subtitles}
\alias{write.subtitles}
\title{Write subtitles}
\usage{
write.subtitles(x, file, format = "srt")
}
\arguments{
\item{x}{an object of class \code{Subtitles}.}

\item{file}{a character string naming a file for writing.}

\item{format}{a character string giving the file format.
Not used (only SubRip format is currently implemented).}
}
\description{
This function writes a \code{Subtitles} object in a file.
}

